Created more realistic gravity that also reacts to other forces.

+++ src/main/objects/SquareObject.java
@@ -6,6 +6,6 @@ import main.shapes.Polygon2D;
 
 public class SquareObject extends Object {
     public SquareObject() {
-        super(new Polygon2D(4 , 1, new Point2D.Double(2, 0)));
+        super(new Polygon2D(4 , 1, new Point2D.Double(2, 5)));
     }
 }
